+++ plugin/fuzzer/testdata/stage_1/a_void_func.c
@@ -1,3 +1,0 @@
-#include "a_void_func.h"
-
-void f() {}