+++ plugin/fuzzer/testdata/stage_1/a_void_func.h
@@ -1,6 +1,0 @@
-#ifndef A_VOID_FUNC_H
-#define A_VOID_FUNC_H
-
-void f();
-
-#endif // A_VOID_FUNC_H