+++ plugin/fuzzer/testdata/stage_2/int_param_nested_if.hpp
@@ -1,6 +1,0 @@
-#ifndef INT_PARAM_NESTED_IF_HPP
-#define INT_PARAM_NESTED_IF_HPP
-
-void func(int v);
-
-#endif // INT_PARAM_NESTED_IF_HPP